requirements.txt
fastapi
pydantic
requests
streamlit
uvicorn
sqlalchemy
psycopg[binary]
python-dotenv
